+++ classes/client.js
@@ -8,8 +8,6 @@ const PermLevels = require("./permLevels.js");
  /* Will Change this later */
 const Config = require("./Configuration Types/Config.js");
 
-require("./Extendables.js");
-
 const defaultPermStructure = new PermLevels()
   .addLevel(0, false, () => true)
   .addLevel(2, false, (client, msg) => {
@@ -45,6 +43,7 @@ module.exports = class Komada extends Discord.Client {
       finalizers: config.disabled.finalizers || [],
       monitors: config.disabled.monitors || [],
       providers: config.disabled.providers || [],
+      extendables: config.disabled.extendables || [],
     };
     this.funcs = new Loader(this);
     this.argResolver = new ArgResolver(this);
@@ -88,7 +87,8 @@ module.exports = class Komada extends Discord.Client {
   }
 
   validatePermStructure() {
-    const permStructure = this.config.permStructure || defaultPermStructure.structure;
+    const structure = this.config.permStructure instanceof PermLevels ? this.config.permStructure.structure : null;
+    const permStructure = structure || this.config.permStructure || defaultPermStructure.structure;
     if (!Array.isArray(permStructure)) throw "PermStructure must be an array.";
     if (permStructure.some(perm => typeof perm !== "object" || typeof perm.check !== "function" || typeof perm.break !== "boolean")) {
       throw "Perms must be an object with a check function and a break boolean.";