+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - A helper class for generating permission levels. It can be accessed via require('komada').PermLevels, and is used via permlevels.addLevel(level, break, checkFunction). Once you have all levels added, simply pass permlevels.structure to your client.config as the "permStructure" property.
 
 ### Changed
+- Dropped support for fs-extra-promise in favor to fs-extra ^3.0.0.
 - Changed fetchMessages to fetchMessage (backend change)
 - Backend is now class based. Users main files will need to be updated. The interface is the same as creating a discord.js client, only using komada, and with komada config. No more use of start, but client.login(token) is needed now.
 - Usage will no longer be calculated everytime a command is run, but instead stored in command.usage.
@@ -26,7 +27,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Changed core log func, into an event. You can now log anything by running client.emit('log', data, type);
 - runMessageMonitors has been moved into the new Message core event.
 - dotenv dependancy has been changed to a peerdep.
-- fs-extra-promise has been updated to the latest version.
 - Minimum node version is now v7.6.x
 - Remaining Utils have been moved to the classes folder.
 - Use Discord.Permissions to generate and keep cached an implied permissions object, instead of generating a new object every time a command is run.